requirements.txt
# Core API
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Bot Dependencies
DrissionPage>=4.0.0
requests>=2.28.0

# Telegram Notifications
python-telegram-bot>=20.0

# Scheduling (for standalone bot mode)
schedule>=1.2.0
